+++ src/core/prototypes.h
@@ -6,7 +6,7 @@ extern "C" void ActionList(struct ActionTable ** Actions, int * Size);
 extern "C" ERR DeleteFile(const std::string_view & Path, FUNCTION * Callback);
 extern "C" CSTRING ResolveClassID(CLASSID ID);
 extern "C" int AllocateID(IDTYPE Type);
-extern "C" ERR AllocMemory(int Size, MEM Flags, APTR * Address);
+extern "C" ERR AllocMemory(int64_t Size, MEM Flags, APTR * Address);
 extern "C" ERR AccessObject(OBJECTID Object, int MilliSeconds, OBJECTPTR * Result);
 extern "C" ERR CheckAction(OBJECTPTR Object, AC Action);
 extern "C" ERR CheckMemoryExists(MEMORYID ID);
